pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-intel"
version = "0.2.0"
description = "Scrape Reddit, extract painpoints via LLM, build a self-maintaining category taxonomy."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Viktar Dubovik" },
    { name = "Daniil Yurshevich" },
    { name = "Daniil Zabauski" },
]
keywords = ["ai", "llm", "mcp", "reddit", "insight", "claude", "openai", "openclaw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet",
]
dependencies = [
    "httpx>=0.28.0",
    "openai>=1.66.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "sqlite-vec>=0.1.0",
]

[project.optional-dependencies]
demo = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]
mcp = [
    "fastmcp>=3.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage   = "https://github.com/openshrug/reddit-intel"
Repository = "https://github.com/openshrug/reddit-intel"
Issues     = "https://github.com/openshrug/reddit-intel/issues"

[project.scripts]
reddit-intel     = "main:main"
reddit-intel-mcp = "mcp_server:run"

# Flat-layout packaging: engine's modules live at repo root rather than
# inside a single top-level package dir. Explicit py-modules + packages
# list is required because setuptools' auto-discovery would otherwise
# pick up tests/, scrape_dumps/, etc. as packages.
#
# Data files (taxonomy.yaml, config.json, db/schema.sql) are resolved at
# runtime via Path(__file__).parent, which works for editable installs
# (pip install -e .). For a built wheel, MANIFEST.in / package-data would
# be needed — add that when non-editable installs become a real use case.
[tool.setuptools]
py-modules = [
    "main",
    "reddit_scraper",
    "subreddit_pipeline",
    "promoter",
    "category_worker",
    "llm",
    "subriff_scraper",
    "mcp_server",
]
packages = [
    "db",
    "painpoint_extraction",
    "evaluation",
    "evaluation.agentic_eval",
    "evaluation.painpoints_eval",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["scrape_dumps", "demo_cache", ".venv", "docs/_internal"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

# Per-file ignores for legitimate patterns ruff over-flags:
# - tests/ files set up sys.path / load .env before importing app modules
#   (E402), bind locals as in-test documentation (F841), and use compact
#   one-liners in setup blocks (E702).
# - db/category_events.py and llm.py interleave a logger init between
#   imports as a deliberate stylistic choice; not worth churning.
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E402", "F841", "E702"]
"db/category_events.py" = ["E402"]
"llm.py" = ["E402"]

# Default test discovery skips tests/live — those tests hit real OpenAI /
# Reddit APIs, take 15+ minutes total, cost money, and need credentials.
# Run them with `pytest -m live` (or `pytest tests/live/` for ad-hoc runs).
# The `live` marker is auto-applied to everything under tests/live/ via
# tests/live/conftest.py; no per-file decorator is needed.
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-m 'not live'"
markers = [
    "live: requires REDDIT/OPENAI credentials and hits real APIs (skipped by default)",
    "slow: takes >10s",
]
